Add basic project layout with linters
In the scope of this task we need to add and configure following linters:
    - mypy
    - flake8
    - pylint

Since we're going to use flake8 and pylint together, there may be
conflicts between their configurations. Now it's difficult to
track all options and resolve conflicts between them, so we
decided to fix them ad-hoc in cases where it will be needed.

Some files like `requirements.txt` and `whitelist.txt` are left
blank for now, but we hope they're going to be filled soon, so
they shouldn't be considered as redundant. If these files keep
to be blank for a long time - feel free to remove them.

GitHub issues: closes #2.