6: Add page with users from database
After we've created a registration page (issue https://github.com/EugeneTurkin/forum123/issues/5), the next step could be adding functionality for it. Like form validation and creating users in database. But before that it's going to be easier if we already will have an ability to check all registered users on forum. With such feature we can easily make sure that our registration works fine. So we need to add the ability to see registered users before we start to implement registration.

So in the scope of this task we need to add basic, let's say it "index page" where will be listed all users we have in database.

PostgreSQL:
- add `docker-compose.yml` file which sets up `PostgreSQL` database using docker containers

SQLAlchemy:
- add `sqlalchemy` to `requirements.txt` and setup it locally
- add `database.py` which initializes `sqlalchemy` by creating `engine`, `session` and `Base`

Alembic:
- add `alembic` to `requirements.txt` and setup it locally
- initialize and configure `alembic` for our project using `alembic init alembic` command inside project root directory
- connect `alembic` to `PostgreSQL` using `sqlalchemy.url` option inside `alembic.ini`
- connect `alembic` to `sqlalchemy`'s metadata using `target_metadata` variable inside `alembic`'s `env.py`

Models:
- add `models.py` with a simple `User` model with fields: `id`, `username`
- register this model in `alembic`'s `env.py` via importing it below the import of `Base`
- create a migration for `User` model using `alembic revision --autogenerate -m "add users table"`
- apply created migration using `alembic upgrade <revision>`
- add two users to database using  `sqlalchemy`'s model `User` in python interactive shell
- check that users have been added successfully using another query via `sqlalchemy` in python interactive shell

Route:
- update implementation of "index" page to make it show all users from database
for example it's going to be a page with something like this:
> Welcome to forum123!
>
> We already have 2 users joined us:
>     - John
>     - Bob

Other:
- update `README.md` with information about how to run database container and apply migrations to it