19: Add cleanup workflows to CI
We are using reabase strategy for git branches and each feature branch at the end of the work is rebased at the last version of develop branch and contains only one commit with description of what has been done for this feature. When developers work with their feature branches they push a lot of "work in progress" commits, and each push triggers GitHub workflow run. But since all of these commits will be rewritten at the end we want to keep our workflow runs list clean and remove previous runs that were triggered by commits that don't exist after rebase.

Also we want to clean up old workflow runs for deleted branches. There is no sense to keep these runs, because they are related to nonexistent commits from nonexistent branches. More than that, since runs of this workflow will remain after each deleted branch we want to clean them up as well. It means that we will always have only one run of this workflow - the run for the most recent deleted branch.

In the scope of this task we need to add:
    - workflow which removes runs for nonexistent commits
    - workflow which removes runs for nonexistent branches and older runs of itself