33: Run linters in parallel on CI
After we've added docker layers caching (#31), we can go further and improve performance of the CI, by running linters in a parallel way.

In this task we need to add ability to run linters in parallel jobs. Also we want to remove some duplication which will happen, because each linter job will work with docker layers caching - and we need to move this into separate reusable action.

Steps to do:

    - rename `ci.yml` -> `lint.yml`
    - add a verbose name for the wokflow using `name` attribute of the workflow
    - add a verbose name for the run using `run-name` attribute of the workflow
    - create separate job for each linter, which will use buildx cache to build a linter and run it after that
    - add a verbose name for each job  using `name` attribute of the job
    - move docker buildx cache logic into separate action
    - move updating docker buildx cache logic into separate action

After this task is done next step can be improving linters performance using their cache options. See `.mypy_cache` for mypy and `PYLINTHOME` for pylint.