35: Add verbosity to cleanup workflow
Now our cleanup workflows has no `name` and their names in "Actions" tab looks like ".github/workflows/cleanup-runs-for-deleted-branch.yml". More than that we have two cleanup workflows, but it's unnecessary. It would be more readable if we had on "Cleanup" workflow in "Actions" tab.

In the scope of this task we need to move two cleanup workflows into one workflow with two jobs. Add a verbose name to wokflow and workflow run.

Steps to do:

    - merge two cleanup workflows into one with two jobs running conditionally
    - add a verbose name to workflow and its run
    - add a verbose name to each job