43: Add rebase gitflow checks to CI
Since we work with git via rebase strategy we want to prevent non-rebased branches to be merged into develop. To achieve that we can add a new GitHub workflow which will check whether branch is rebased on last develop or not.

In the scope of this task we need to following checks to our CI:

    - branch is rebased
    - commits are squashed

Steps to do:

    - write a python script `envs/ci/scripts/branches/lint.py` which runs checks and prints errors
      it should also return 0 in case of no errors and 1 otherwise
    - add a new job to `.github/workflows/lint.yml` which runs our script