23: Add topics list for logged in users
After we've finished with [Basic "Sign In" / "Sign Up"](https://github.com/EugeneTurkin/forum123/milestone/1?closed=1), we can go further and work with forum's core functionality - topics and posts.

Before we will add an ability to create topics, we want to be able to see already created topics. It will help us with testing when we will work on topics creation functionality.

So in the scope of this task we need to add an ability to see topics from database at `/topics` for logged in users.

Steps to do:

    - add model `Topic` with fields: `id`, `author_id`, `created_at`, `description`, `title`
    - create migrations and apply them to database
    - add two topics to database using python interactive shell
    - check that topics have been added successfully using another query in python interactive shell
    - add `/topics` page and make it show all topics from database for logged in users
      for each topic let's display its title, description, author's _name_, and date of creation
    - make topics ordered by creation date (recent first)
    - add redirect to `/topics` after successful sign in