24: Add topic creation page
After we've created `Topic` model and added an ability to see created topics (#23), we can go further and start implementing topics creation.

In the scope of this task we need add to basic topic creation page without any functionality. We just need to see creation form when we go to `/topics/create`)

Steps to do:

    - add html template with topic creation form in it (form fields should be: `title`, `description`)
    - add `Cancel` button, which will clear form data (idk is it required or not) and redirect to `/topics`
    - add flask route for topic creation page `/topics/create` which renders this form
    - add "Create topic" button to the `/topics` page template and make it point to `/topics/create`