26: Add topic page
After we've added topics creation (#25) and ability to see list of created topics (#23), we can go further and add a page for topic itself.

In the scope of this task we need to add an ability to see topic content at `/topics/{topic_id}`. For now it will only be author, title, description and creation date (no posts listed for now)

Steps to do:

    - add a template for topic page
    - add a router for `/topics/{topic_id}` which will get topic from database and return correct template rendered
    - add an ability to go to the appropriate topic page by the link for each topic in `/topics`