76: Remove unused topic check in if-statement
While working on error handling for topic routes (#55) we used a walrus operator `:=` to check that there is no such topic in database.
After that we got this code:
```python
if not (topic := session.query(Topic).filter_by(id=topic_id).first()):
    return ...
if topic and form.validate_on_submit():
    ...
```
Actulally there is not sense to check for topic presence in the second if-statement, since if there is no topic, the first if-statement will just exit the function at the return statement.

So in the scope of this task we need to remove a check for topic presence in the second if statement.