92: Fix models registration for alembic
While working on separating users, topics and posts related functionality (#77, #80, #81). We forgot to update our alembic configuration file `migrations/env.py`. This file imports models, to let alembic know which models we have, before running the migrations.

So in the scope of this task we need to update models import in `migrations/env.py`.