101: Add REST endpoint to get topic
We are going to add REST API interface to our forum, and we decided to begin with creating endpoints without authentication. It won't be secure but it will make development process easier. We will add authentication to these endpoints in the future.

We need to add an endpoint which will return the info about particular topic. It will be `author_id`, `created_at`, `description`, and `title`.

So in the scope of this task we need to create an endpoint which will return topic info in json format, or 404 if topic was not found.