139: Add missing pytest-env to requirements
While working on adding unit-testing layout (#58), we used `env = ` section in `pytest.ini` config file to provide a correct postgres port for our testing database. But this section works only due-to [pytest-env](https://pypi.org/project/pytest-env/) plugin. We installed this plugin locally, but forgot to add it to the `requirements-dev.txt` so at the time of merging this functionality it seems to be worked, but it actually not.

This was revealed after some time. We noticed that `pytest` cannot connect to database for some reason. And at that time we solved this problem by adding `POSTGRES_PORT=5433` before `pytest` command in "Run tests" section of our README.md (#130). But to solve this problem we could just add pytest-env to our requirements, and leave `pytest` command without any additional env vars.

So in the scope of this task we need to revert changes made for (#130), and add `pytest-env` to our `requirements-dev.txt` file.