141: Fix pylint configuration to check tests
While working on adding unit-testing layout (#58) we forgot to setup pylint to check `tests` directory as well.

So in the scope of this task we need to fix pylint configuration and update `README.md` file to make pylint check `tests` directory.

***

Since pylint doesn't have anything like `per-file-ignores` we had to use [pylint-per-file-ignores](https://pypi.org/project/pylint-per-file-ignores/) plugin.

Since this plugin cannot work only with configuration in `pyproject.toml` we had to add one to define `per-file-ignores` in it.

***

All of this doesn't look good  so it's a temporary solution and I hope we will be able to find a better way to setup per file ignores for `pylint` in the future.