134: Fix POSTGRES_HOST default value in config
While working on adjusting coverage to check only rest code (#129) we mistakenly merged another default value for `POSTGRES_HOST` in `config.py`. This wasn't intentional and the previous value was correct.

So in the scope of this task we need to change `POSTGRES_HOST` default value in config to the previous value - `localhost`.