161: Fix coverage exclude rules
Since we want to cover only REST API related functionality, we need to exclude some additional code related to SSR part of the application. This work is related to #129.

So in the scope of this ticket we need to exclude code in `src/shared/reqparse.py` and `src/user/utils.py`. As an outcome of this work we want to see 100% coverage, because actually all REST API code is covered at this moment.