154: Separate parser instances for each route
Now we create one parser instance for all routes within a file. And after that each route adds each own arguments to the parser.

This means that when we run different routes in our application each route within a file will try to parse arguments added by different routes. But we need a parser to have arguments only for one route each time.

So in the scope of this ticket we need to fix that by creating a separate parser instance for each route.