153: Fix response for get post REST endpoint
While working on adding REST endpoint to get post (#102), we added wrong response body:
```python
return {
    "author_id": post.id,
    "created_at": str(post.created_at),
    "body": post.body,
    "topic_id": post.topic_id,
}
```
This body doesn't contain "id" and has a wrong value in "author_id".

So in the scope of this task we need to fix these two fields. "id" should contain a post id and "author_id" should contain id of the post author.